refactor dirs and separate exporter into base and version specific
